requirements.txt
numpy
pandas
scikit-learn
tensorflow
keras
streamlit
matplotlib
pillow